+++ app/client/src/components/editorComponents/DynamicHeightOverlay/ui/AutoHeightLimitHandleBorder.tsx
@@ -15,11 +15,7 @@ const AutoHeightLimitHandleBorder = styled.div<
   );
   background-size: 8% 1px;
   background-repeat: repeat-x;
-  // position: absolute;
-  // left: 0;
-  // right: 0;
   height: 1px;
-  // top: 0px;
   width: 100%;
 
   ${(props) => (props.isActive ? `background-color: ${OVERLAY_COLOR}` : "")}
